+++ src/components/common/slider/DualSlider.tsx
