+++ src/components/common/slider/FlavorSlider.tsx
@@ -1,6 +1,6 @@
 import { Slider } from "@/components/ui/slider";
 import { cn } from "@/lib/utils";
-import { Badge } from "../ui/badge";
+import { Badge } from "../../ui/badge";
 
 interface FlavorSliderProps {
   value: number;
